+++ tests/InstallDynamicsRole.ps1
@@ -60,7 +60,6 @@ try {
         Write-Host "Creating Dynamics admin db on $dbHostName";
         Install-Dynamics365Server `
             -MediaDir C:\Install\Dynamics\Dynamics365Server90RTMEnu `
-            -LicenseKey KKNV2-4YYK8-D8HWD-GDRMW-29YTW `
             -CreateDatabase `
             -ServerRoles BackEnd, DeploymentAdministration `
             -SqlServer $dbHostName\SQLInstance01 `